BERT_api/requirements.txt
# Function dependencies, for example:
# package>=version
absl-py==0.12.0
alembic==1.7.1
astunparse==1.6.3
autopage==0.4.0
cachetools==4.2.2
cmaes==0.8.2
cmd2==2.2.0
dataclasses==0.6
dill==0.3.4
entrypoints==0.3
farasapy==0.0.14
filelock==3.0.12
flatbuffers
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
importlib-resources==5.2.0
Keras-Preprocessing==1.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.4.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.3
protobuf==3.17.3
PyArabic==0.6.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyperclip==1.8.2
pyreadline3==3.3
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
sacremoses==0.0.45
scipy==1.7.1
stevedore==3.4.0
termcolor==1.1.0
terminado==0.9.4
threadpoolctl==2.2.0
tokenizers==0.9.4
torch==1.8.1
torchvision==0.9.1
transformers==4.2.1
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
google-cloud-storage==1.42.3
flask==2.0.2